test for Validation.Target warnings
